backend/requirements.txt
boto3==1.13.19
django-cors-headers==3.3.0
django-environ==0.4.5
django-extensions~=2.2.9
django-filter==2.2.0
Django==3.0.6
djangorestframework==3.11.0
psycopg2==2.8.5
requests==2.23.0
urllib3==1.25.9
